refactor(window-state): split window_state.rs into per-phase submodules
Replace src-tauri/src/window_state.rs (679 LOC) with a window_state/
directory that separates the schema, migration, file I/O, monitor
matching, and the restore/save lifecycle phases:

- schema.rs            - WindowState, MonitorSnapshot, LogicalRect,
                         PersistedStore, CURRENT_VERSION, default_normal
- migration.rs         - v0->v1 physical->logical migration + tests
- persistence.rs       - state_file_path, load_store, save_store +
                         JSON round-trip test
- monitor_matching.rs  - exact_match / intersects / nearest_by_center
                         three-tier fallback
- restore.rs           - restore_on_setup, restoreable_state + the
                         clamp / cross-DPI / fallback tests
- save.rs              - schedule_save (debounced) + save_now (sync
                         flush on close)
- mod.rs               - module decls, WindowStateStore + impl,
                         test_support helpers, pub use re-exports

Public surface unchanged. The four lib.rs call sites (line 78 .manage,
line 82 schedule_save, line 88 save_now, line 200 restore_on_setup)
keep resolving through window_state::* re-exports.
